Add missing javadoc + rename 'Provide Bundle' transaction prefix
