+++ src/main/java/org/ehrbase/fhirbridge/fhir/bundle/ProvideBundleAuditStrategy.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2020-2021 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.ehrbase.fhirbridge.fhir.bundle;
 
 import org.openehealth.ipf.commons.audit.AuditContext;
@@ -6,10 +22,16 @@ import org.openehealth.ipf.commons.ihe.fhir.audit.FhirAuditDataset;
 import org.openehealth.ipf.commons.ihe.fhir.audit.FhirAuditStrategy;
 import org.openehealth.ipf.commons.ihe.fhir.support.OperationOutcomeOperations;
 
+/**
+ * Implementation of {@link org.openehealth.ipf.commons.ihe.core.atna.AuditStrategy AuditStrategy}
+ * for 'Provide Bundle' transaction.
+ *
+ * @since 1.0.0
+ */
 public class ProvideBundleAuditStrategy extends FhirAuditStrategy<FhirAuditDataset> {
 
-    public ProvideBundleAuditStrategy(boolean serverSide) {
-        super(serverSide, OperationOutcomeOperations.INSTANCE);
+    public ProvideBundleAuditStrategy() {
+        super(true, OperationOutcomeOperations.INSTANCE);
     }
 
     @Override
