+++ src/main/java/org/ehrbase/fhirbridge/fhir/bundle/converter/DiagnosticReportLabConverter.java
@@ -1,6 +1,5 @@
-package org.ehrbase.fhirbridge.fhir.bundle;
+package org.ehrbase.fhirbridge.fhir.bundle.converter;
 
-import org.ehrbase.fhirbridge.fhir.bundle.converter.AbstractBundleConverter;
 import org.ehrbase.fhirbridge.fhir.common.Profile;
 import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.DiagnosticReport;
