+++ src/main/resources/META-INF/services/org/apache/camel/component/bundle-provide
